# openapi.json
{
    "openapi": "3.0.3",
    "info": {
        "info": "Cereals",
        "version": "3.0"
    },
    "paths": {
        "/all": {
            "get": {
                "summary": "Get all cereals from collection",
                "description": "Returns all rows with all the cereals",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                ],
                "responses": {
                    "200": {
                        "description": "Fetch successfull",
                        "schema": {
                            "$ref": "#/definitions/schemas/cereal"
                        }
                    }
                }
            }
        },
        "/one?id={id}": {
            "get": {
                "summary": "Get one cereal from collection by ID",
                "description": "Returns one row of the cereal with the provided ID",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the resource",
                        "required": true,
                        "type": "integer",
                        "format": "int"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Fetch successfull",
                        "schema": {
                            "$ref": "#/definitions/schemas/cereal"
                        }
                    },
                    "400": {
                        "description": "No ID input given"
                    },
                    "404": {
                        "description": "No cereal with requested ID"
                    }
                }
            }
        }, 
        "/calories": {
            "get": {
                "summary": "Get all cereals with their calories",
                "description": "Get all cereals with their calories ordered ascending",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                ],
                "responses": {
                    "200": {
                        "description": "Fetch successfull",
                        "schema": {
                            "$ref": "#/definitions/schemas/calories"
                        }
                    }
                }
            }
        },
        "/price": {
            "get": {
                "summary": "Get all cereals with their prices",
                "description": "Get all cereals with their prices ordered ascending",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                ],
                "responses": {
                    "200": {
                        "description": "Fetch successfull",
                        "schema": {
                            "$ref": "#/definitions/schemas/price"
                        }
                    }
                }
            }
        },
        "/country": {
            "get": {
                "summary": "Get all cereals with selected store country",
                "description": "Get all cereals with selected store country",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        
                        "name": "country",
                        "in": "path",
                        "description": "ISO 3166 alpha-2 code of the wanted country",
                        "required": true,
                        "type": "string",
                        "format": "str"
                        
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Fetch successfull",
                        "schema": {
                            "$ref": "#/definitions/schemas/country"
                        }
                    },
                    "400": {
                        "description": "No country input given"
                    },
                    "404": {
                        "description": "Bad request"
                    }
                }
            }
        },
        "/specification": {
            "get": {
                "summary": "Get OpenAPI specification",
                "description": "Returns OpenAPI specificaiton",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                ],
                "responses": {
                    "200": {
                        "description": "Fetch successfull",
                        "schema": {
                            "$ref": "application/json"
                        }
                    }
                }
            }
        },
        "/postone": {
            "post": {
                "summary": "Insert one row to the collection",
                "description": "Insert one row to the collection",
                "produces": [
                    "application/json"
                ],
                "requestBody": {
                    "content": {
                        "schema": {
                            "$ref": "#/definitions/schemas/cerealId"
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Insert successfull",
                        "schema": {
                            "$ref": "#/definitions/schemas/cerealId"
                        }
                    },
                    "400": {
                        "description": "Variable not correct"
                    },
                    "406": {
                        "description": "Missing valid attribute names"
                    }
                }
            }
        },
        "/putone?id={id}": {
            "put": {
                "summary": "Updates one row in the collection",
                "description": "Updates one row in the collection",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the resource",
                        "required": false,
                        "type": "integer",
                        "format": "int"
                    }
                ],
                "requestBody": {
                    "content": {
                        "schema": {
                            "$ref": "#/definitions/Resource"
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Update successfull"
                    },
                    "404": {
                        "description": "Cereal with this ID doesn't exist"
                    },
                    "406": {
                        "description": "Missing valid attribute names"
                    }
                }
            }
        },
        "/deleteone?id={id}": {
            "delete": {
                "summary": "Deletes one row in the collection",
                "description": "Deletes one row in the collection passed by ID",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the resource",
                        "required": false,
                        "type": "integer",
                        "format": "int"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Delete successfull"
                    },
                    "400": {
                        "description": "No ID input given"
                    },
                    "404": {
                        "description": "No cereal with given ID"
                    }
                }
            }
        }
    },
    "definitions": {
        "schemas": {
            "cereal": {
                "id_cereal": {
                    "type": "integer"
                  },
                  "cereal_name": {
                    "type": "string"
                  },
                  "type": {
                    "type": "string"
                  },
                  "number_of_calories": {
                    "type": "integer"
                  },
                  "price": {
                    "type": "float"
                  },
                  "id_store": {
                    "type": "integer"
                  },
                  "id_manufacturer": {
                    "type": "integer"
                  },
                  "ingredients": {
                    "type": "array"
                  }
            },
            "cerealId": {
                  "cereal_name": {
                    "type": "string"
                  },
                  "type": {
                    "type": "string"
                  },
                  "number_of_calories": {
                    "type": "integer"
                  },
                  "price": {
                    "type": "float"
                  },
                  "id_store": {
                    "type": "integer"
                  },
                  "id_manufacturer": {
                    "type": "integer"
                  },
                  "ingredients": {
                    "type": "array"
                  }
            },
            "calories": {
                "cereal_name": {
                    "type": "string"
                  },
                  "number_of_calories": {
                    "type": "integer"
                  }
            },
            "price": {
                "cereal_name": {
                    "type": "string"
                  },
                  "price": {
                    "type": "float"
                  }
            },
            "country": {
                "id_store": {
                    "type": "integer"
                  },
                  "id_cereal": {
                    "type": "integer"
                  },
                  "cereal_name": {
                    "type": "string"
                  },
                  "id_manufacturer": {
                    "type": "integer"
                  },
                  "type": {
                    "type": "string"
                  },
                  "number_of_calories": {
                    "type": "integer"
                  },
                  "price": {
                    "type": "float"
                  },
                  "ingredients": {
                    "type": "array"
                  },
                  "store_name": {
                    "type": "string"
                  },
                  "store_country_iso": {
                    "type": "string"
                  }
            }
        }
    }

}